pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "data-prep"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "pydantic>=2.12.4",
]

[project.optional-dependencies]
full = [
    "numpy>=2.0.0",
    "dotenv>=0.9.9",
    "langchain>=1.0.8",
    "langchain-core>=1.0.7",
    "langchain-openai>=1.0.3",
    "librosa>=0.11.0",
    "openai-whisper>=20250625",
    "pytube>=15.0.0",
    "pytubefix>=10.3.4",
    "scipy>=1.16.3",
    "soundfile>=0.13.1",
    "tensorflow==2.18.0 ; sys_platform == 'darwin'",
    "tensorflow-hub>=0.16.1",
    "tensorflow-metal==1.2.0 ; sys_platform == 'darwin'",
    "tensorflow[and-cuda]>=2.18.0 ; sys_platform != 'darwin'",
    "torch>=2.9.1",
    "torchaudio>=2.9.1",
    "torchcodec>=0.8.1",
    "tqdm>=4.67.1",
    "yt-dlp>=2025.11.12",
]

cpu = ["torch==2.9.1", "torchaudio==2.9.1"]
cu126 = ["torch==2.9.1", "torchaudio==2.9.1"]
cu128 = ["torch==2.9.1", "torchaudio==2.9.1"]
cu130 = ["torch==2.9.1", "torchaudio==2.9.1"]

[tool.setuptools.packages.find]
where = ["src"]
include = ["sca_data*"]

[tool.uv]
conflicts = [[{ extra = "cpu" }, { extra = "cu126" }, { extra = "cu128" }, { extra = "cu130" }]]

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu", extra = "cpu" },
    { index = "pytorch-cu126", extra = "cu126" },
    { index = "pytorch-cu128", extra = "cu128" },
    { index = "pytorch-cu130", extra = "cu130" },
]
torchaudio = [
    { index = "pytorch-cpu", extra = "cpu" },
    { index = "pytorch-cu126", extra = "cu126" },
    { index = "pytorch-cu128", extra = "cu128" },
    { index = "pytorch-cu130", extra = "cu130" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[dependency-groups]
dev = [
    "notebook>=7.5.0",
]
